workflow-scripts/requirements.txt
pytest
GitPython
